change!: rename crps_uniqueness to uniqueness change!: make intra_hd accept vector, matrix and array feat: add reliability as complementary of intra_hd chore: fix problems with font in svg logo
The main change of this commit is to intra_hd.
Now, intra_hd accepts a vector, a 2D matrix and a 3D array.
It returns directly the reliability of the vector, of  each CRP,
and of each CRP across all devices respectively.

The `reliability` function has been added to calculate the
complementary value of intra_hd, as it's supposed to be.

This commit also fixes a problem with the font in the logo.svg file.