change!: remove crps_weight in favor of uniformity and bitaliasing feat: add metrics class to compute all metrics docs: add Readme.Rmd refactor: move metrics to formulas.R
This commit provides the new pufmetrics class to compute all metrics automatically.
This also allows automatic exploration of the metrics.

Also, the `crps_weight` function has been deleted in favour of uniformity and bitaliasing
since the previous commit removed parallel code. The code should be fast enough to not require
parallelisation. If need, it should be implemented by the user